+++ Store.Application/Store.Application.csproj
@@ -10,4 +10,13 @@
       <Folder Include="Services\" />
     </ItemGroup>
 
+    <ItemGroup>
+      <PackageReference Include="AutoMapper" Version="14.0.0" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Store.Domain\Store.Domain.csproj" />
+    </ItemGroup>
+
 </Project>
