+++ Store.Domain/Entities/ProductInfo.cs
@@ -1,0 +1,44 @@
+using System;
+using Store.Domain.Abstractions;
+using Store.Domain.Enums;
+
+namespace Store.Domain.Entities;
+
+public class ProductInfo : BaseEntity
+{
+    /// <summary>
+    /// Название товара
+    /// </summary>
+    public string Title { get; set; }
+
+    /// <summary>
+    /// Цена товара
+    /// </summary>
+    public decimal Price { get; set; }
+
+    /// <summary>
+    /// Описание товара
+    /// </summary>
+    public string Description { get; set; }
+
+    /// <summary>
+    /// Статус товара
+    /// </summary>
+    public ProductStatus Status { get; set; }
+
+    /// <summary>
+    /// Путь к изображению товара
+    /// </summary>
+    public string ImagePath { get; set; }
+
+    /// <summary>
+    /// процент скидки
+    /// </summary>
+    public decimal Discount { get; set; }
+
+    public Guid CategoryId { get; set; }
+    public ProductCategory Category { get; set; }
+
+    public Guid? SubCategoryId { get; set; }
+    public ProductSubCategory SubCategory { get; set; }
+}
