+++ Store.Domain/Entities/ProductSubCategory.cs
@@ -1,12 +1,19 @@
 using System;
 using System.Collections.Generic;
+using Store.Domain.Abstractions;
 
-namespace Store.Data.Entities;
+namespace Store.Domain.Entities;
 
 public class ProductSubCategory : BaseEntity
 {
+    /// <summary>
+    /// Название подкатегории
+    /// </summary>
     public string SubCategoryName { get; set; }
 
+    /// <summary>
+    /// Путь к изображению подкатегории
+    /// </summary>
     public string ImagePath { get; set; }
 
     public Guid CategoryId { get; set; }
